requirements.txt
asttokens==2.4.1
colorama==0.4.6
comm==0.2.1
debugpy==1.8.1
decorator==5.1.1
docx2pdf==0.1.8
et-xmlfile==1.1.0
executing==2.0.1
ipykernel==6.29.2
ipython==8.21.0
ipywidgets==8.1.2
jedi==0.19.1
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyterlab_widgets==3.0.10
lxml==5.1.0
matplotlib-inline==0.1.6
nest-asyncio==1.6.0
numpy==1.26.4
openpyxl==3.1.2
packaging==23.2
pandas==2.2.0
parso==0.8.3
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
Pygments==2.17.2
PyPDF2==3.0.1
python-dateutil==2.8.2
python-docx==1.1.0
pytz==2024.1
pywin32==306
pyzmq==25.1.2
six==1.16.0
stack-data==0.6.3
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
typing_extensions==4.10.0
tzdata==2024.1
wcwidth==0.2.13
widgetsnbextension==4.0.10
